requirements.txt
numpy==1.26.4
pandas==2.2.2
psycopg2-binary
python-dotenv
scikit-learn